+++ tests/test_md/forms.py
@@ -1,0 +1,9 @@
+from django.forms import ModelForm
+
+from .models import MDPlaceModel
+
+
+class MDPlaceForm(ModelForm):
+    class Meta:
+        model = MDPlaceModel
+        fields = ('idno', 'company_type_1', 'company_type_2', 'license_plate')