+++ docs/changelog.rst
@@ -118,6 +118,9 @@ New flavors:
 - Added local flavor for Cuba
   (`gh-292 <https://github.com/django/django-localflavor/pull/292>`_).
 
+- Added local flavor for Moldova
+  (`gh-309 <https://github.com/django/django-localflavor/pull/309>`_)
+
 New fields for existing flavors:
 
 - Added KWAreaSelect form field